change type

+++ modules/koki-portal/src/main/kotlin/com/wutsi/koki/portal/listing/form/ListingForm.kt
@@ -47,7 +47,7 @@ data class ListingForm(
     val leaseTerm: Int? = null,
     val noticePeriod: Int? = null,
     val advanceRent: Int? = null,
-    val securityDeposit: Long? = null,
+    val securityDeposit: Int? = null,
     val currency: String? = null,
     val sellerContactId: Long? = null,
     val sellerAgentCommission: Double? = null,
